// src/app/models/players/UserModalData.ts
export class  UserModalData
{
  result:string;
  userId: string;
  userName: string;
  firstName: string;
  middleName: string;
  lastName: string;
  dob: string;
  address: string;
  address2: string;
  city: string;
  pin: string;
  state: string;
  gender: string;
  email: string;
  mobile: string;
  kycVerified: string;
  registrationDate: string;
  referralCode: string;
  isLocked : string;
  deposit: string;
  isFraud:string;
  fraudReason:string;
  markedFraudDate:string;
  markedNonFraudDate:string;
  withdrawable: string;
  promoBonus: string;
  lockedBonus: string;
  totalCashGamesPlayed: string;
  totalFreeGamesPlayed: string;
  totalRake: string;
  totalDepositAmount: string;
  totalWithdrawalAmount: string;
  pokerDeposit : string;
  pokerWithdrawal: string;
  code : string;
}